# examples/taxiBackend/taxi-openapi.json
{"openapi":"3.1.0","info":{"title":"Tiny Taxi Backend","version":"0.1.0"},"paths":{"/riders":{"get":{"summary":"Riders","operationId":"riders_riders_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/RiderOut"},"type":"array","title":"Response Riders Riders Get"}}}}}},"post":{"summary":"Create Rider","operationId":"create_rider_riders_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiderCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiderOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drivers":{"get":{"summary":"Drivers","operationId":"drivers_drivers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DriverOut"},"type":"array","title":"Response Drivers Drivers Get"}}}}}},"post":{"summary":"Create Driver","operationId":"create_driver_drivers_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DriverCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DriverOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips":{"get":{"summary":"Trips","operationId":"trips_trips_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TripOut"},"type":"array","title":"Response Trips Trips Get"}}}}}},"post":{"summary":"Request Trip","operationId":"request_trip_trips_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}":{"get":{"summary":"Trip Detail","operationId":"trip_detail_trips__trip_id__get","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"integer","title":"Trip Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}/assign":{"post":{"summary":"Assign","operationId":"assign_trips__trip_id__assign_post","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"integer","title":"Trip Id"}},{"name":"driver_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Driver Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}/ping":{"post":{"summary":"Ping","operationId":"ping_trips__trip_id__ping_post","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"integer","title":"Trip Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripProgress"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}/simulate":{"post":{"summary":"Simulate","operationId":"simulate_trips__trip_id__simulate_post","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"integer","title":"Trip Id"}},{"name":"duration","in":"query","required":false,"schema":{"type":"integer","default":60,"title":"Duration"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripProgress"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trips/{trip_id}/{action}":{"post":{"summary":"Change Status","operationId":"change_status_trips__trip_id___action__post","parameters":[{"name":"trip_id","in":"path","required":true,"schema":{"type":"integer","title":"Trip Id"}},{"name":"action","in":"path","required":true,"schema":{"$ref":"#/components/schemas/TripStatus"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard.json":{"get":{"summary":"Dashboard Json","operationId":"dashboard_json_dashboard_json_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard":{"get":{"summary":"Dashboard","operationId":"dashboard_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DriverCreate":{"properties":{"name":{"type":"string","title":"Name"},"car_plate":{"type":"string","title":"Car Plate"}},"type":"object","required":["name","car_plate"],"title":"DriverCreate"},"DriverOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"car_plate":{"type":"string","title":"Car Plate"}},"type":"object","required":["id","name","car_plate"],"title":"DriverOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RiderCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"RiderCreate"},"RiderOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"RiderOut"},"TripCreate":{"properties":{"rider_id":{"type":"integer","title":"Rider Id"},"origin":{"type":"string","title":"Origin"},"dest":{"type":"string","title":"Dest"}},"type":"object","required":["rider_id","origin","dest"],"title":"TripCreate"},"TripOut":{"properties":{"id":{"type":"integer","title":"Id"},"rider_id":{"type":"integer","title":"Rider Id"},"driver_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Driver Id"},"origin":{"type":"string","title":"Origin"},"dest":{"type":"string","title":"Dest"},"fare":{"type":"number","title":"Fare"},"remaining_km":{"type":"number","title":"Remaining Km"},"status":{"$ref":"#/components/schemas/TripStatus"}},"type":"object","required":["id","rider_id","driver_id","origin","dest","fare","remaining_km","status"],"title":"TripOut"},"TripProgress":{"properties":{"id":{"type":"integer","title":"Id"},"remaining_km":{"type":"number","title":"Remaining Km"},"status":{"$ref":"#/components/schemas/TripStatus"},"complete":{"type":"boolean","title":"Complete"}},"type":"object","required":["id","remaining_km","status","complete"],"title":"TripProgress"},"TripStatus":{"type":"string","enum":["requested","accepted","started","ended"],"title":"TripStatus"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}